requirements.txt
socket
alabaster
altgraph
arrow
astroid
asttokens
atomicwrites==1.4.0
attrs
autopep8
Babel
backcall
bcrypt
beautifulsoup4
binaryornot
black
bleach
Brotli
certifi==2021.5.30
cffi
chardet
charset-normalizer
click
cloudpickle
colorama
comm
contourpy==1.2.0
cookiecutter
cryptography
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
debugpy
decorator
defusedxml
diff-match-patch
dill
docstring-to-markdown
docutils
executing
Faker==21.0.0
fastjsonschema
flake8
flit_core
fonttools==4.48.1
future
gmpy2
gnupg==2.3.1
idna==2.10
imagesize
importlib-metadata
inflection==0.5.1
intervaltree
ipykernel
ipython
ipython-genutils
isort
jaraco.classes
jedi
jellyfish
Jinja2==3.0.1
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab-pygments
keyring
kiwisolver==1.4.5
lazy-object-proxy
markdown-it-py
MarkupSafe==2.0.1
matplotlib==3.8.2
matplotlib-inline
mccabe
mdurl
mistune
more-itertools
mpmath
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
numpy==1.26.2
numpydoc
oracledb==2.0.0
packaging
pandas==2.1.3
pandocfilters
paramiko
parso
pathspec
pefile
pexpect
pickleshare
Pillow==9.4.0
platformdirs
pluggy
ply==3.11
prompt-toolkit
psutil
ptyprocess
pure-eval
pycodestyle
pycparser
pycryptodome
pydocstyle
pyflakes
Pygments
pyinstaller
pyinstaller-hooks-contrib
pylint
pylint-venv
pyls-spyder==0.4.0
PyNaCl
pyOpenSSL
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip
PyQtWebEngine==5.15.6
PySocks
python-dateutil
python-lsp-black
python-lsp-jsonrpc==1.0.0
python-lsp-server
python-slugify
pytz
pywin32==305.1
pywin32-ctypes
PyYAML
pyzmq
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
referencing
requests==2.25.1
rich
rope
rpds-py
Rtree
sip
six
snowballstemmer
sortedcontainers
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
stack-data
sympy
text-unidecode
textdistance
three-merge
tinycss2
toml
tomli
tomlkit
tornado
traitlets
typing_extensions
tzdata==2023.3
ujson
Unidecode
urllib3==1.26.6
watchdog
wcwidth
webencodings==0.5.1
whatthepatch
win-inet-pton
wrapt
yapf
zipp
